requirements.txt
Django==5.1.4
djangorestframework==3.15.2
requests==2.32.3
djangorestframework_simplejwt==5.4.0
mysqlclient>=2.2.7
drf_yasg==1.21.8